VIITE-1947 Moved change info related stuff to kmtk-package.
