+++ digiroad2-oracle/src/main/scala/fi/liikennevirasto/digiroad2/DummySerializer.scala
@@ -2,7 +2,7 @@ package fi.liikennevirasto.digiroad2
 
 import java.io.File
 
-import fi.liikennevirasto.digiroad2.client.vvh.ChangeInfo
+import fi.liikennevirasto.digiroad2.client.kmtk.ChangeInfo
 import fi.liikennevirasto.digiroad2.linearasset.RoadLink
 import fi.liikennevirasto.digiroad2.util.RoadLinkSerializer
