+++ digiroad2-oracle/src/test/scala/fi/liikennevirasto/digiroad2/ChangeTypeSpec.scala
@@ -1,6 +1,6 @@
 package fi.liikennevirasto.digiroad2
 
-import fi.liikennevirasto.digiroad2.client.vvh.{ChangeInfo, ChangeType}
+import fi.liikennevirasto.digiroad2.client.kmtk.{ChangeInfo, ChangeType}
 import org.scalatest.{FunSuite, Matchers}
 
 /**
