+++ digiroad2-viite/src/main/scala/fi/liikennevirasto/viite/process/RoadAddressChangeInfoMapper.scala
@@ -1,8 +1,8 @@
 package fi.liikennevirasto.viite.process
 
 import fi.liikennevirasto.digiroad2.Point
-import fi.liikennevirasto.digiroad2.client.vvh.ChangeType._
-import fi.liikennevirasto.digiroad2.client.vvh.ChangeInfo
+import fi.liikennevirasto.digiroad2.client.kmtk.ChangeInfo
+import fi.liikennevirasto.digiroad2.client.kmtk.ChangeType._
 import org.slf4j.LoggerFactory
